feat: new controller_build_info metric
One of the metrics #12589.

Add new, static metric which contains labels from the product version
fields.

This is common good practice, and puts the information about which
version of a controller was running that a given point in time.

This can be useful when diagnosing issues that occurred, and also
shows the progress of a version rollout when that happens.

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>